refactor: simplify receiver card component and update address service implementation
