+++ app/pubspec.yaml
@@ -53,6 +53,7 @@ dependencies:
   http: ^1.5.0
   firebase_app_check: ^0.3.2+10
   dart_jsonwebtoken: ^3.2.0
+  infinite_scroll_pagination: ^5.1.1
 
 dev_dependencies:
   flutter_test:
